Basement Workshops now dynamically link to parent Settlement containers
When a basement door is placed, the workshop in the basement is
dynamically linked to whatever settlement the door was placed in.  This
unfortunately does not include items from Provisioner-linked
settlements, as that seems to be handled by the engine.

Other minor changes to the ESP (version number, removing unnecessary
floors in the Military Bunker Epsilon)